Add files via upload
Achieved an accuracy of 84.8% with a multiple linear regression model to predict future sales, identifying TV as the most significant predictor. Analyzed the dataset to examine the impact of TV, Radio, and Newspapers on sales . Developed the model by visualizing relationships with scatter plots, assessing correlations with a heat map, and splitting the data for accurate predictions.